+++ src/wininput.c
@@ -585,6 +585,14 @@ void
 //    alt_fn ? W("Alt+F12") : ct_sh ? W("Ctrl+Shift+S") : null
 //  );
 
+  uint status_line = term.st_type == 1 ? MF_CHECKED
+                   : term.st_type == 0 ? MF_UNCHECKED
+                   : MF_GRAYED;
+  //__ Context menu:
+  modify_menu(ctxmenu, IDM_STATUSLINE, status_line, _W("Status Line"),
+    null
+  );
+
   uint options_enabled = config_wnd ? MF_GRAYED : MF_ENABLED;
   EnableMenuItem(ctxmenu, IDM_OPTIONS, options_enabled);
   EnableMenuItem(sysmenu, IDM_OPTIONS, options_enabled);
@@ -737,6 +745,7 @@ win_init_ctxmenu(bool extended_menu, bool with_user_commands)
   AppendMenuW(ctxmenu, MF_ENABLED | MF_UNCHECKED, IDM_SCROLLBAR, 0);
   AppendMenuW(ctxmenu, MF_ENABLED | MF_UNCHECKED, IDM_FULLSCREEN_ZOOM, 0);
 //  AppendMenuW(ctxmenu, MF_ENABLED | MF_UNCHECKED, IDM_FLIPSCREEN, 0);
+  AppendMenuW(ctxmenu, MF_ENABLED | MF_UNCHECKED, IDM_STATUSLINE, 0);
   AppendMenuW(ctxmenu, MF_SEPARATOR, 0, 0);
   if (extended_menu) {
     //__ Context menu: generate a TTY BRK condition (tty line interrupt)
@@ -1422,13 +1431,13 @@ toggle_dim_margins(struct term* term_p)
 }
 
 void
-toggle_status_line(struct term* term_p)
+(toggle_status_line)(struct term* term_p)
 {
   TERM_VAR_REF(true)
     
   if (term.st_type == 1)
     term_set_status_type(0, 0);
-  else
+  else if (term.st_type == 0)
     term_set_status_type(1, 0);
 }
 
@@ -1810,7 +1819,9 @@ static uint
 {
   TERM_VAR_REF(true)
     
-  return term.st_type == 1 ? MF_CHECKED : MF_UNCHECKED;
+  return term.st_type == 1 ? MF_CHECKED
+       : term.st_type == 0 ? MF_UNCHECKED
+       : MF_GRAYED;
 }
 
 static uint